+++ src/petigapce.c
@@ -62,7 +62,7 @@ static PetscErrorCode PCSetUp_EBE_CreateMatrix(Mat A,Mat *B)
           ierr = MatSetSizes(mat,m,n,M,N);CHKERRQ(ierr);
           ierr = MatSetBlockSize(mat,bs);CHKERRQ(ierr);
 
-          ierr = PetscMalloc1(ia[na],&newja);CHKERRQ(ierr);
+          ierr = PetscMalloc1((size_t)ia[na],&newja);CHKERRQ(ierr);
           for (j=0; j<ia[na]; j++) newja[j] = ja[j] + cstart;
           if (aij)   {ierr = MatMPIAIJSetPreallocationCSR(mat,ia,newja,NULL);CHKERRQ(ierr);}
           if (baij)  {ierr = MatMPIBAIJSetPreallocationCSR(mat,bs,ia,newja,NULL);CHKERRQ(ierr);}
@@ -143,13 +143,13 @@ static PetscErrorCode PCSetUp_EBE(PC pc)
 
     n = nen*dof;
     ierr = PetscBLASIntCast(n,&m);CHKERRQ(ierr);
-    ierr = PetscMalloc1(n,&indices);CHKERRQ(ierr);
-    ierr = PetscMalloc1(n*n,&values);CHKERRQ(ierr);
-    ierr = PetscMalloc1(m,&ipiv);CHKERRQ(ierr);
+    ierr = PetscMalloc1((size_t)n,&indices);CHKERRQ(ierr);
+    ierr = PetscMalloc1((size_t)n*(size_t)n,&values);CHKERRQ(ierr);
+    ierr = PetscMalloc1((size_t)m,&ipiv);CHKERRQ(ierr);
     lwork = -1; work = &lwkopt;
     LAPACKgetri_(&m,values,&m,ipiv,work,&lwork,&info);
     lwork = (info==0) ? (PetscBLASInt)work[0] : m*128;
-    ierr = PetscMalloc1(lwork,&work);CHKERRQ(ierr);
+    ierr = PetscMalloc1((size_t)lwork,&work);CHKERRQ(ierr);
 
     ierr = IGABeginElement(iga,&element);CHKERRQ(ierr);
     while (IGANextElement(iga,element)) {
