+++ test/IGAGeometryMap.c
@@ -459,8 +459,8 @@ PetscErrorCode IGAComputeScalarFull(IGA iga,Vec vecU,
   PetscValidScalarPointer(S,3);
   IGACheckSetUp(iga,1);
 
-  ierr = PetscCalloc1(n,&localS);CHKERRQ(ierr);
-  ierr = PetscMalloc1(n,&workS);CHKERRQ(ierr);
+  ierr = PetscCalloc1((size_t)n,&localS);CHKERRQ(ierr);
+  ierr = PetscMalloc1((size_t)n,&workS);CHKERRQ(ierr);
 
   /* Get local vector U and array */
   ierr = IGAGetLocalVecArray(iga,vecU,&localU,&arrayU);CHKERRQ(ierr);
@@ -476,7 +476,7 @@ PetscErrorCode IGAComputeScalarFull(IGA iga,Vec vecU,
       /* Quadrature loop */
       ierr = IGAElementBeginPoint(element,&point);CHKERRQ(ierr);
       while (IGAElementNextPoint(element,point)) {
-        ierr = PetscMemzero(workS,n*sizeof(PetscScalar));CHKERRQ(ierr);
+        ierr = PetscMemzero(workS,(size_t)n*sizeof(PetscScalar));CHKERRQ(ierr);
         ierr = Scalar(point,U,n,workS,ctx);CHKERRQ(ierr);
         ierr = IGAPointAddArray(point,n,workS,localS);CHKERRQ(ierr);
       }
@@ -539,8 +539,8 @@ int main(int argc, char *argv[]) {
 
   ierr = IGASetGeometryDim(iga,dim);CHKERRQ(ierr);
   iga->rational = PETSC_TRUE;
-  ierr = PetscMalloc1(3*3*(dim-1),&iga->rationalW);CHKERRQ(ierr);
-  ierr = PetscMalloc1(3*3*(dim-1)*dim,&iga->geometryX);CHKERRQ(ierr);
+  ierr = PetscMalloc1(3*3*(size_t)(dim-1),&iga->rationalW);CHKERRQ(ierr);
+  ierr = PetscMalloc1(3*3*(size_t)((dim-1)*dim),&iga->geometryX);CHKERRQ(ierr);
   {
     PetscInt i,j,k,m=dim-1;
     PetscInt posx=0,posw=0;
